-rename Channel to Deferred -polling for remote promises
git-svn-id: https://svn.code.sf.net/p/waterken/code/server/trunk@1727 ec340d7e-3b38-0410-a49c-f1b72a8843da

+++ waterken/ref_send/src/org/ref_send/list/List.java
@@ -43,12 +43,12 @@ List<T> implements Iterable<T>, Serializable {
     /**
      * link count
      */
-    private int capacity;
+    private long capacity;
 
     /**
      * element count
      */
-    private int size;
+    private long size;
 
     private
     List() {
@@ -107,7 +107,7 @@ List<T> implements Iterable<T>, Serializable {
     /**
      * Gets the element count.
      */
-    public int
+    public long
     getSize() { return size; }
 
     /**
